refactoring config
